improvement: move query and key to their own file in platformLibrary directory
