--- ui/src/services/k8s/volumes.key.js
@@ -1,5 +1,0 @@
-export const volumesKey = {
-  all: ['volumes', 'list'],
-  volumeObject: ['volumes', 'object'],
-  persitant: ['volumes', 'Persistant'],
-};
