--- ui/src/services/prometheus/prometheus.key.js
@@ -1,4 +1,0 @@
-export const prometheusKey = {
-  query: (query) => ['query', query],
-  queryRange: ['queryRange'],
-};
